refactor: Movie DTO 수정 - 인기영화 순으로 출력을 위해 popularity field추가 및 float형 데이터 타입변경 발생
DataInitController를 만들어서 Json파일에 있는 영화 리스트들을 DB에 옮겼다.
그 과정에서 Float형으로 class type을 casting할 수 없다는 오류가 발생하여 해당 DTO들의 field를 바꿔주게 됐다.
score, popularity => double type으로 변경

MovieResponseListDTO
list출력 시 popularity 기준으로 정렬을 하여 출력할 필요가 있었기에 해당 popularity field를 entity에서 가져오게 만들었다.
imageUrl도 가져오게 했다. 인기영화의 포스트를 보여주기 위해서였다.

+++ community/src/main/java/com/movie/community/controller/dto/request/MovieRequestDTO.java
@@ -14,12 +14,12 @@ public class MovieRequestDTO {
 	private Long id;
 	private String title;
 	private String description;
-	private float score;
+	private double score;
 	private int voteCount;
 	private String imageUrl;
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
 	private LocalDate releasedDate;
-	private float popularity;
+	private double popularity;
 	private String genres;
 	private String holiday;
